refactor: move MMD approach to own module.
