+++ rul_adapt/approach/__init__.py
@@ -1,4 +1,5 @@
-from .tbigru import MmdApproach, select_features, VibrationFeatureExtractor
+from .tbigru import select_features, VibrationFeatureExtractor
+from .mmd import MmdApproach
 from .adarul import AdaRulApproachPretraining, AdaRulApproach
 from .consistency import ConsistencyApproachPretraining, ConsistencyApproach
 from .dann import DannApproach
